works in notification click

+++ lib/screens/mainScreen/payment.dart
@@ -90,15 +90,15 @@ class _PaymentState extends State<Payment> {
     }
   }
 
-  addThisOrderToFirestore() async {
+  Future<void> addThisOrderToFirestore() async {
     double total =
         double.parse(widget.totalAfterTax) + double.parse(widget.delvierCost);
-    Firestore.instance.collection('order').add({
-      'payment': '',
-      'driverID': id,
-      'driverName': name,
+    await Firestore.instance.collection('order').add({
+      'payment': '100',
+      'driverID': '',
+      'driverName': '',
       'orderID': orderID,
-      'date': '',
+      'date': DateTime.now().toString(),
       'status': '0',
       'address': widget.address,
       'city': widget.city,
@@ -119,7 +119,6 @@ class _PaymentState extends State<Payment> {
           ? iosDeviceInfo.identifierForVendor
           : androidInfo.androidId,
     });
-    paymantPage();
   }
 
   bool paymentMethodForRiyadh = false;
@@ -204,7 +203,7 @@ class _PaymentState extends State<Payment> {
         await Geocoder.local.findAddressesFromCoordinates(coordinates);
     var first = addresses.first;
     city = first.locality;
-    if (city == "Riyadh") {
+    if (city == "Riyadh" || city == "الرياض") {
       setState(() {
         paymentMethodForRiyadh = true;
       });
@@ -216,7 +215,7 @@ class _PaymentState extends State<Payment> {
     isoCode = code.alpha3;
 
     setState(() {});
-    addThisOrderToFirestore();
+    paymantPage();
   }
 
   List<ItemShow> cart = [];
@@ -278,6 +277,7 @@ class _PaymentState extends State<Payment> {
     fetchMyCart();
   }
 
+  bool addFirestore = false;
   @override
   Widget build(BuildContext context) {
     return Scaffold(
@@ -355,8 +355,8 @@ class _PaymentState extends State<Payment> {
                               double.parse(widget.delvierCost);
                           Firestore.instance.collection('order').add({
                             'payment': 'cash',
-                            'driverID': id,
-                            'driverName': name,
+                            'driverID': '',
+                            'driverName': '',
                             'orderID': orderID,
                             'date': DateTime.now().toString(),
                             'status': '0',
@@ -434,37 +434,44 @@ class _PaymentState extends State<Payment> {
                   javascriptMode: JavascriptMode.unrestricted,
                   onPageFinished: (url) async {
                     if (url == "http://geniusloop.co/payment/succed.php") {
-                      FocusScope.of(context).requestFocus(FocusNode());
-                      twilioFlutter.sendSMS(
-                          toNumber: phone,
-                          messageBody:
-                              'الوان ولمسات\nمرحبا ${widget.firstName} لقد تم أستلام طلبك\nرقم طلبك هو $orderID');
-                      sendEmailToCustomer();
-                      addCartToast(word("Succssful", context));
-                      navIndex = 3;
-                      DBHelper.deleteAllItem("cart");
-                      Navigator.popUntil(context, (route) => route.isFirst);
+                      if (!addFirestore) {
+                        addThisOrderToFirestore().then((value) {
+                          FocusScope.of(context).requestFocus(FocusNode());
+                          twilioFlutter.sendSMS(
+                              toNumber: phone,
+                              messageBody:
+                                  'الوان ولمسات\nمرحبا ${widget.firstName} لقد تم أستلام طلبك\nرقم طلبك هو $orderID');
+                          sendEmailToCustomer();
+                          addCartToast(word("Succssful", context));
+                          navIndex = 3;
+                          DBHelper.deleteAllItem("cart");
+                          Navigator.popUntil(context, (route) => route.isFirst);
+                        });
+                        setState(() {
+                          addFirestore = true;
+                        });
+                      }
                     } else if (url ==
                         "http://geniusloop.co/payment/failed.php") {
                       FocusScope.of(context).requestFocus(FocusNode());
                       errorToast(word("Failed", context));
-                      Navigator.push(
-                        context,
-                        MaterialPageRoute(
-                          builder: (context) => Payment(
-                            totalAfterTax: widget.totalAfterTax,
-                            price: widget.price,
-                            buyPrice: widget.buyPrice,
-                            onThemeChanged: widget.onThemeChanged,
-                            changeLangauge: widget.changeLangauge,
-                            firstName: widget.firstName,
-                            lastName: widget.lastName,
-                            phone: widget.phone,
-                            lat: widget.lat,
-                            long: widget.long,
-                          ),
-                        ),
-                      );
+                      // Navigator.push(
+                      //   context,
+                      //   MaterialPageRoute(
+                      //     builder: (context) => Payment(
+                      //       totalAfterTax: widget.totalAfterTax,
+                      //       price: widget.price,
+                      //       buyPrice: widget.buyPrice,
+                      //       onThemeChanged: widget.onThemeChanged,
+                      //       changeLangauge: widget.changeLangauge,
+                      //       firstName: widget.firstName,
+                      //       lastName: widget.lastName,
+                      //       phone: widget.phone,
+                      //       lat: widget.lat,
+                      //       long: widget.long,
+                      //     ),
+                      //   ),
+                      // );
                     }
                   },
                 ),
@@ -482,12 +489,7 @@ class _PaymentState extends State<Payment> {
     final message = mail.Message()
       ..from = mail.Address(username, 'Your name')
       ..recipients.add(widget.email) //recipent email
-      // ..ccRecipients.addAll([
-      //   'destCc1@example.com',
-      //   'destCc2@example.com'
-      // ]) //cc Recipents emails
-      // ..bccRecipients.add(Address(
-      //     'bccAddress@example.com')) //bcc Recipents emails
+
       ..subject = word("subject_email", context) //subject of the email
       ..html =
           '<h1>${word('body_email', context)} </h1><h1 style="color:pink">$orderID</h1> <br> <table cellpadding="0" cellspacing="0" class="sc-gPEVay eQYmiW" style="vertical-align: -webkit-baseline-middle; font-size: medium; font-family: Arial;"><tbody><tr><td><table cellpadding="0" cellspacing="0" class="sc-gPEVay eQYmiW" style="vertical-align: -webkit-baseline-middle; font-size: medium; font-family: Arial;"><tbody><tr><td style="padding: 0px; vertical-align: middle;"><h3 color="#000000" class="sc-fBuWsC eeihxG" style="margin: 0px; font-size: 18px; color: rgb(0, 0, 0);"><span>Yousef</span><span>&nbsp;</span><span>Alyousef</span></h3><p color="#000000" font-size="medium" class="sc-fMiknA bxZCMx" style="margin: 0px; color: rgb(0, 0, 0); font-size: 14px; line-height: 22px;"><span>Supplier</span></p><p color="#000000" font-size="medium" class="sc-dVhcbM fghLuF" style="margin: 0px; font-weight: 500; color: rgb(0, 0, 0); font-size: 14px; line-height: 22px;"><span>Marketing</span><span>&nbsp;|&nbsp;</span><span>Rfoof Store</span></p><table cellpadding="0" cellspacing="0" class="sc-gPEVay eQYmiW" style="vertical-align: -webkit-baseline-middle; font-size: medium; font-family: Arial; width: 100%;"><tbody><tr><td height="30"></td></tr><tr><td color="#F2547D" direction="horizontal" height="1" class="sc-jhAzac hmXDXQ" style="width: 100%; border-bottom: 1px solid rgb(242, 84, 125); border-left: none; display: block;"></td></tr><tr><td height="30"></td></tr></tbody></table><table cellpadding="0" cellspacing="0" class="sc-gPEVay eQYmiW" style="vertical-align: -webkit-baseline-middle; font-size: medium; font-family: Arial;"><tbody><tr height="25" style="vertical-align: middle;"><td width="30" style="vertical-align: middle;"><table cellpadding="0" cellspacing="0" class="sc-gPEVay eQYmiW" style="vertical-align: -webkit-baseline-middle; font-size: medium; font-family: Arial;"><tbody><tr><td style="vertical-align: bottom;"><span color="#F2547D" width="11" class="sc-jlyJG bbyJzT" style="display: block; background-color: rgb(242, 84, 125);"><img src="https://cdn2.hubspot.net/hubfs/53/tools/email-signature-generator/icons/phone-icon-2x.png" color="#F2547D" width="13" class="sc-iRbamj blSEcj" style="display: block; background-color: rgb(242, 84, 125);"></span></td></tr></tbody></table></td><td style="padding: 0px; color: rgb(0, 0, 0);"><a href="tel:+18126827296" color="#000000" class="sc-gipzik iyhjGb" style="text-decoration: none; color: rgb(0, 0, 0); font-size: 12px;"><span>+18126827296</span></a> | <a href="tel:+8126827296" color="#000000" class="sc-gipzik iyhjGb" style="text-decoration: none; color: rgb(0, 0, 0); font-size: 12px;"><span>+8126827296</span></a></td></tr><tr height="25" style="vertical-align: middle;"><td width="30" style="vertical-align: middle;"><table cellpadding="0" cellspacing="0" class="sc-gPEVay eQYmiW" style="vertical-align: -webkit-baseline-middle; font-size: medium; font-family: Arial;"><tbody><tr><td style="vertical-align: bottom;"><span color="#F2547D" width="11" class="sc-jlyJG bbyJzT" style="display: block; background-color: rgb(242, 84, 125);"><img src="https://cdn2.hubspot.net/hubfs/53/tools/email-signature-generator/icons/email-icon-2x.png" color="#F2547D" width="13" class="sc-iRbamj blSEcj" style="display: block; background-color: rgb(242, 84, 125);"></span></td></tr></tbody></table></td><td style="padding: 0px;"><a href="mailto:api.chc1989@gmail.com" color="#000000" class="sc-gipzik iyhjGb" style="text-decoration: none; color: rgb(0, 0, 0); font-size: 12px;"><span>api.chc1989@gmail.com</span></a></td></tr><tr height="25" style="vertical-align: middle;"><td width="30" style="vertical-align: middle;"><table cellpadding="0" cellspacing="0" class="sc-gPEVay eQYmiW" style="vertical-align: -webkit-baseline-middle; font-size: medium; font-family: Arial;"><tbody><tr><td style="vertical-align: bottom;"><span color="#F2547D" width="11" class="sc-jlyJG bbyJzT" style="display: block; background-color: rgb(242, 84, 125);"><img src="https://cdn2.hubspot.net/hubfs/53/tools/email-signature-generator/icons/link-icon-2x.png" color="#F2547D" width="13" class="sc-iRbamj blSEcj" style="display: block; background-color: rgb(242, 84, 125);"></span></td></tr></tbody></table></td><td style="padding: 0px;"><a href="http://geniusloop.co/" color="#000000" class="sc-gipzik iyhjGb" style="text-decoration: none; color: rgb(0, 0, 0); font-size: 12px;"><span>http://geniusloop.co/</span></a></td></tr></tbody></table><table cellpadding="0" cellspacing="0" class="sc-gPEVay eQYmiW" style="vertical-align: -webkit-baseline-middle; font-size: medium; font-family: Arial;"><tbody><tr><td height="30"></td></tr></tbody></table><a href="https://www.hubspot.com/email-signature-generator?utm_source=create-signature" target="_blank" rel="noopener noreferrer" class="sc-gisBJw kDlVKO" style="font-size: 12px; display: block; color: rgb(0, 0, 0);"></a></td></tr></tbody></table></td></tr></tbody></table>';
@@ -499,71 +501,7 @@ class _PaymentState extends State<Payment> {
     } on mail.MailerException catch (e) {
       print('Message not sent. \n' +
           e.toString()); //print if the email is not sent
-      // e.toString() will show why the email is not sending
     }
-  }
-
-  double min = 0.0;
-  String name = '';
-  String id = '';
-  double calculateDistance(lat1, lon1, lat2, lon2) {
-    var p = 0.017453292519943295;
-    var c = cos;
-    var a = 0.5 -
-        c((lat2 - lat1) * p) / 2 +
-        c(lat1 * p) * c(lat2 * p) * (1 - c((lon2 - lon1) * p)) / 2;
-    return 12742 * asin(sqrt(a));
-  }
-
-  getTheDeriver() async {
-    try {
-      double customerLat = double.parse(widget.lat);
-      double customerLong = double.parse(widget.long);
-      List<DriverModel> driverList = new List();
-      await Firestore.instance
-          .collection('employee')
-          .getDocuments()
-          .then((value) {
-        value.documents.forEach((e) {
-          if (e['accept'] == '1') {
-            double deriverLat = e['lat'];
-            double deriverLong = e['long'];
-            List<dynamic> data = [
-              {
-                "lat": customerLat,
-                "lng": customerLong,
-              },
-              {
-                "lat": deriverLat,
-                "lng": deriverLong,
-              }
-            ];
-            double totalDistance = 0;
-            for (var i = 0; i < data.length - 1; i++) {
-              totalDistance += calculateDistance(data[i]["lat"], data[i]["lng"],
-                  data[i + 1]["lat"], data[i + 1]["lng"]);
-            }
-            driverList.add(DriverModel(
-                name: e['name'], id: e['id'], distance: totalDistance));
-          }
-        });
-
-        for (var i = 0; i < driverList.length; i++) {
-          if (min == 0.0) {
-            min = driverList[i].distance;
-            name = driverList[i].name;
-            id = driverList[i].id;
-          } else {
-            if (min > driverList[i].distance) {
-              min = driverList[i].distance;
-              name = driverList[i].name;
-              id = driverList[i].id;
-            }
-          }
-        }
-        setState(() {});
-      });
-    } catch (e) {}
   }
 }
 
@@ -576,12 +514,4 @@ paymentToast(String text) {
       backgroundColor: Colors.green,
       textColor: Colors.white,
       fontSize: 16.0);
-}
-
-class DriverModel {
-  final String name;
-  final String id;
-  final double distance;
-
-  DriverModel({this.name, this.id, this.distance});
 }
